DeleteBookUseCaseに存在しないIDを指定した場合の処理実装。

+++ src/main/java/com/example/examination2/application/DeleteBookUseCase.java
@@ -1,5 +1,6 @@
 package com.example.examination2.application;
 
+import com.example.examination2.application.exception.BookNotFoundException;
 import com.example.examination2.domain.repository.BookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,6 +11,9 @@ public class DeleteBookUseCase {
     private final BookRepository bookRepository;
 
     public void deleteBook(String id) {
+        if (bookRepository.getBookById(id).isEmpty()) {
+            throw new BookNotFoundException(id);
+        }
         bookRepository.deleteBook(id);
     }
 }
